+++ tests/rag_system/test_async_no_encode_rag_system.py
@@ -184,7 +184,7 @@ def test_bridging_no_encode_rag_system(
         _bridge_version = "0.1.0"
         _bridge_extra = "my-bridge"
         _framework = "my-bridge-framework"
-        _compatible_versions = ["0.1.x"]
+        _compatible_versions = {"min": "0.1.1", "max": "0.2.0"}
         _method_name = "to_bridge"
 
         def to_bridge(self) -> None:
@@ -211,6 +211,9 @@ def test_bridging_no_encode_rag_system(
         rag_system.bridges["my-bridge-framework"]
         == _TestBridgeMixin.get_bridge_metadata()
     )
+
+    # cleanup
+    del BridgedAsyncNoEncodeRAGSystem.bridges[_TestBridgeMixin._framework]
 
 
 def test_rag_system_to_sync(